+++ CHANGELOG.md
@@ -12,7 +12,7 @@
     - Support custom port for PIO Home server
     - Override default `~/.platformio/penv` folder with environment variable `PLATFORMIO_PENV_DIR`
 * Added new setting `platformio-ide.pioHomeServerHttpPort` which sets a default HTTP port for PIO Home server (issue [#832](https://github.com/platformio/platformio-vscode-ide/issues/832))
-* Added new setting `platformio-ide.disableAutostartPIOHomeServer` which disables automatic starting of PIO Home server (issue [#888](https://github.com/platformio/platformio-vscode-ide/issues/888))
+* Added new setting `disablePIOHomeStartup` which disables showing PIO Home at startup (issue [#888](https://github.com/platformio/platformio-vscode-ide/issues/888))
 * Replaced *.png icons with *.svg (issue [#755](https://github.com/platformio/platformio-vscode-ide/issues/755))
 * Fixed an issue with multiple instances of PIO Home when opening new project (issue [#624](https://github.com/platformio/platformio-vscode-ide/issues/624))